Add fwupd build dependencies, pylint3->pylint
When running "make everything", the last make step ("cab_file")
requires both php and gcab. Without them, the image files still get
made but the build is marked as a failure.

php and gcab are needed to make a build for fwupd in the last make step
so they were added to the build dependencies in a new step.

The pylint3 build dependency was updated to be pylint since most
distributions have pylint3 as an alias for pylint and some (at least
Ubuntu and Pop!_OS) don't have pylint3.