Add link to incoming-call USB disconnections; update UCM2 paths
This commit adds a link to the known workarounds for USB
disconnections that occur on incoming calls. While these are
probably not the fault of the modem fw, a full understanding of
why the workarounds work might reveal details of why the
non-incoming-calls USB disconnections occur (which would be
useful to know, no matter whether this is the fault of the modem
fw or not).

This commit also updates the UCM2 paths from the PP and the PPP,
although the PPP path appears to be missing from upstream
alsa-ucm-conf, so this commit lists it as "Mobian/trixie".  The
old paths retained as "older setups", so that the reader has
enough of a clue that these paths might not yet be stable, and
s/he may need to search around to check the current path and name
of right .conf file to edit. The URLs to the flow diagrams for
Alsa UCM and to the current alsa-ucm-conf upstream repository are
also linked in this commit, to make it easier for people to find
these.